// src/models/product.model.ts
export interface Products{
    id?: string;
    category_id: string;
    user_id: string;
    imgUrl: string;
    name: string;
    brief_description: string;
    description: string;
    units: number;
    measurement?: string;
    price: number;
    lat?: number;
    lng?:number;
    location?: string;
}